Add a program that prints first 50 fabionacci nums fix

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 int i;
-long long n1, n2, fabionacci[50];
+long int n1, n2, fabionacci[50];
 
 n1 = 1;
 n2 = 2;
